ml/requirements.txt
# ============================================================
# ML Pipeline Requirements
# Fake Certificate Detection - GPU-enabled (CUDA 11.8)
# ============================================================

# Deep Learning
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
efficientnet_pytorch==0.7.1

# Vision-Language Model (CLIP)
transformers==4.36.2
open-clip-torch==2.23.0

# OCR
pytesseract==0.3.10
pdf2image==1.16.3

# Image processing
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0

# ML utilities
scikit-learn==1.3.2
numpy==1.26.3
tqdm==4.66.1
matplotlib==3.8.2

# API / IO
requests==2.31.0
aiofiles==23.2.1

# Logging
loguru==0.7.2
